requirements.txt
Django==1.7.1
Fabric==1.10.1
argparse==1.2.1
dj-database-url==0.3.0
djangorestframework==3.0.2
ecdsa==0.11
gunicorn==19.1.1
paramiko==1.15.2
psycopg2==2.5.4
pycrypto==2.6.1
wsgiref==0.1.2
